move view state classes to a separate package
