+++ app/src/main/java/com/shevaalex/android/plugev/presentation/mapscreen/viewstate/FilteringRowViewState.kt
