+++ app/src/main/java/com/shevaalex/android/plugev/presentation/mapscreen/viewstate/MapScreenViewState.kt
@@ -1,8 +1,9 @@
-package com.shevaalex.android.plugev.presentation.mapscreen
+package com.shevaalex.android.plugev.presentation.mapscreen.viewstate
 
 import com.google.android.gms.maps.model.LatLng
 import com.shevaalex.android.plugev.domain.model.ChargingStation
 import com.shevaalex.android.plugev.presentation.common.ui.UiState
+import com.shevaalex.android.plugev.presentation.mapscreen.*
 
 data class MapScreenViewState(
     val cameraPosition: LatLng = LatLng(MAP_DEFAULT_LATITUDE, MAP_DEFAULT_LONGITUDE),
